refactor: merge models view into functions view
