+++ src/screens/surrealist/views/functions/FunctionPropertiesPanel/style.module.scss
