--- src/screens/surrealist/views/models/ModelsPanel/style.module.scss
@@ -1,4 +1,0 @@
-.scroller > div {
-	width: 100%;
-    table-layout: fixed;
-}
